+++ DataTables/src/main/java/me/Fupery/DataTables/Test.java
@@ -1,0 +1,17 @@
+package me.Fupery.DataTables;
+public class Test {
+
+    public static void main(String[] args) {
+        PixelTable table;
+        try {
+            table = DataTables.loadTable(4);
+        } catch (DataTables.InvalidResolutionFactorException e) {
+            e.printStackTrace();
+            return;
+        }
+
+        for (float f :table.getYawBounds()) {
+            System.out.println(f);
+        }
+    }
+}