+++ plugin/src/main/java/me/Fupery/ArtMap/Menu/API/BasicMenu.java
@@ -4,8 +4,6 @@ import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.ClickType;
 import org.bukkit.event.inventory.InventoryType;
 
-import com.github.Fupery.InvMenu.Utils.MenuType;
-
 import me.Fupery.ArtMap.Menu.Event.MenuCloseReason;
 import me.Fupery.ArtMap.Menu.Handler.CacheableMenu;