+++ plugin/src/main/java/me/Fupery/ArtMap/Menu/HelpMenu/ArtistArtworksMenu.java
@@ -5,8 +5,6 @@ import java.util.List;
 import java.util.UUID;
 import java.util.logging.Level;
 
-import com.github.Fupery.InvMenu.Utils.SoundCompat;
-
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
@@ -19,6 +17,7 @@ import me.Fupery.ArtMap.Config.Lang;
 import me.Fupery.ArtMap.IO.MapArt;
 import me.Fupery.ArtMap.Menu.API.ChildMenu;
 import me.Fupery.ArtMap.Menu.API.ListMenu;
+import me.Fupery.ArtMap.Menu.API.SoundCompat;
 import me.Fupery.ArtMap.Menu.Button.Button;
 import me.Fupery.ArtMap.Menu.Event.MenuCloseReason;
 import me.Fupery.ArtMap.Menu.Handler.CacheableMenu;